added user access control

+++ src/routes/article.js
@@ -1,5 +1,6 @@
 const express = require('express');
 const Article = require('../models/article');
+const User = require('../models/user');
 const router = new express.Router();
 const validateSchema = require('../validation/joi');
 
@@ -7,6 +8,7 @@ const validateSchema = require('../validation/joi');
 
 router.post('/articles/add', async (req, res) => {
   const article = new Article(req.body);
+  article.author = req.user._id; //take id of user access control
   try {
     const validationResult = validateSchema.validate(req.body, {
       abortEarly: false,
@@ -21,12 +23,13 @@ router.post('/articles/add', async (req, res) => {
       res.redirect('/');
     }
   } catch (error) {
+    console.log(error);
     return res.render('error', {
       error: 'request denied check after some time',
     });
   }
 });
-router.get('/articles/add', async (req, res) => {
+router.get('/articles/add', ensureAuthenticated, async (req, res) => {
   res.render('add', {
     title: 'Add article',
   });
@@ -49,14 +52,22 @@ router.get('/', async (req, res) => {
 //fetch single article
 router.get('/articles/:id', async (req, res) => {
   const article = await Article.findById(req.params.id);
+
+  const user = await User.findById(article.author);
+  //article.author._id take id 🔼 fetch user using id and set author name
   res.render('article', {
     article: article,
+    author: user.name, //set user name as a author
   });
 });
 
 //for edit articles🔂
-router.get('/articles/edit/:id', async (req, res) => {
+router.get('/articles/edit/:id', ensureAuthenticated, async (req, res) => {
   const article = await Article.findById(req.params.id);
+  //if author not match then can't edit post
+  if (article.author != req.user.id) {
+    return res.redirect('/');
+  }
   res.render('edit_article', {
     article: article,
   });
@@ -65,10 +76,10 @@ router.get('/articles/edit/:id', async (req, res) => {
 // Update Submit POST Route
 router.post('/articles/edit/:id', async (req, res) => {
   const { title, author, body } = req.body;
-  const query = { _id: req.params.id };
+  const id = { _id: req.params.id };
   try {
     const result = await Article.findByIdAndUpdate(
-      query,
+      id,
       { title, author, body },
       { new: true }
     );
@@ -93,6 +104,11 @@ router.post('/articles/edit/:id', async (req, res) => {
 });
 
 router.get('/articles/delete/:id', async (req, res) => {
+  console.log(req.user.id);
+  if (!req.user._id) {
+    //if article id not match then it can't delete post
+    return res.redirect('/');
+  }
   const article = await Article.findByIdAndDelete(req.params.id);
   if (article) {
     res.redirect('/');
@@ -102,4 +118,13 @@ router.get('/articles/delete/:id', async (req, res) => {
     });
   }
 });
+
+// Access Control || for protect any route
+function ensureAuthenticated(req, res, next) {
+  if (req.isAuthenticated()) {
+    return next();
+  } else {
+    res.redirect('/user/login');
+  }
+}
 module.exports = router;
